+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+
+- **Breaking:** Replace `env` with `setenv` or `setenv_if_unset`, depending on whether inherited values may override it.
+- **Breaking:** Rename `unset_env` to `unsetenv`.
+
 ## v0.1.0 (2026-03-30)
 
 First release of Sheld. See README for more information.